requirements.txt
Flask==2.3.2
gunicorn==20.1.0
Werkzeug==2.3.7
python-dotenv